+++ CHANGELOG.rst
@@ -700,7 +700,7 @@ Changelog
 
 * New function ``mp_batch(foldername, glob_str)`` for parallel processing of a folder of list files.
 
-0.8.0 (2018-02-XX)
+0.8.0 (2018-03-19)
 -----------------
 
 * Added ``recursive`` and ``n_proc`` keywords to ``mp_batch``, and changed return type to ``None``.
@@ -718,3 +718,8 @@ Changelog
 * Added option to run PySight with a predetermined config file: ``main.run(cfg_file='/path/to/file.json')``.
 
 * New integration tests.
+
+0.8.1 (2018-03-19)
+------------------
+
+* Bug fix in ``setup.py``.